feat: enhance category management with caching and validation checks

+++ src/main/java/com/fat/BUS/Services/CategoryService.java
@@ -30,7 +30,11 @@ public class CategoryService implements ICategoryService {
 
     @Override
     public List<CategoryViewDTO> getAllCategories() {
-        return categoryDAO.getAll();
+        if(categoriesCache.isEmpty()){
+            List<CategoryViewDTO> data = categoryDAO.getAll();
+            categoriesCache.addAll(data);
+        }
+        return categoriesCache;
     }
 
     @Override
@@ -41,23 +45,58 @@ public class CategoryService implements ICategoryService {
 
     @Override
     public List<CategoryViewDTO> filterCategoryByList(String keyword) {
-        // TODO: Implement filter from ArrayList
-        return null;
+        if (keyword == null || keyword.trim().isEmpty())
+            return getAllCategories();
+        String searchKey = keyword.trim().toLowerCase();
+        return categoriesCache.stream()
+            .filter(c -> c.getName().toLowerCase().contains(searchKey))
+            .collect(Collectors.toList());
     }
 
     @Override
     public void createCategory(CreateOrUpdateCategoryDTO dto) {
-        categoryDAO.add(dto);
+        String name = dto.getName();
+        if(name == null || name.trim().isEmpty())
+            throw new RuntimeException("Tên danh mục không được để trống");
+        boolean isExist = categoriesCache.stream()
+        .anyMatch(c -> c.getName().equalsIgnoreCase(name));
+        //duyệt qua từng cái để kiểm tra tên trùng
+        if(isExist){
+            throw new RuntimeException("Tên danh mục đã tồn tại");
+        }
+        Integer newId = categoryDAO.add(dto);
+        CategoryViewDTO newCategory = new CategoryViewDTO(newId, name);
+        categoriesCache.add(0, newCategory);
     }
 
     @Override
     public void updateCategory(CreateOrUpdateCategoryDTO dto) {
+        Integer id = dto.getId();
+        String name = dto.getName();
+        if(name == null || name.trim().isEmpty())
+            throw new RuntimeException("Tên danh mục không được để trống");
+        boolean isExist = categoriesCache.stream()
+        .anyMatch(c -> !c.getId().equals(id) && c.getName().equalsIgnoreCase(name));
+        if (isExist) {
+            throw new RuntimeException("Tên danh mục đã tồn tại");
+        }
+
         categoryDAO.update(dto);
+        for (int i = 0; i < categoriesCache.size(); i++) {
+            if (categoriesCache.get(i).getId().equals(id)) {
+                categoriesCache.set(i, new CategoryViewDTO(id, name));
+                break;
+            }
+        }
     }
 
     @Override
     public void deleteCategory(Integer id) {
+        if (categoryDAO.isExistAnyProductInCategory(id)) {
+            throw new RuntimeException("Không thể xóa danh mục vì đang có sản phẩm thuộc danh mục này");
+        }
         categoryDAO.delete(id);
+        categoriesCache.removeIf(c -> c.getId().equals(id));
     }
 }
